Adds an initialization-time option to use an existing screen buffer.
This change allows users the option of supplying an existing screen buffer to use.

This breaks the internal encapsulation of the screen buffer, which may be a bad thing for any sort of public library, so I’m not sure if this is a good addition to the general-purpose release, but it fits my own needs.